get_collectors

Get collectors for a SurveyMonkey survey

get_collectors(
  survey_id,
  page = 1,
  per_page = NULL,
  sort_by = NULL,
  sort_order = NULL,
  name = NULL,
  start_date = NULL,
  end_date = NULL,
  include = NULL,
  all_pages = TRUE,
  oauth_token = get_token()
)

Arguments

survey_id

The id of the survey whose collectors you want, find it with browse_surveys.

page

Integer number to select which page of resources to return. By default is 1.

per_page

Number of resources to return per page.

sort_by

Column to sort by. Can be one of: id, date_modified, type, status, name.

sort_order

Sort data by the value in `sort_by`. Can be one of: ASC, DESC.

name

Nickname of collector to search against

start_date

Collectors must be created after this date. Date string in format YYYY-MM-DDTHH:MM:SS (no offset)

end_date

Collectors must be created before this date. Date string in format YYYY-MM-DDTHH:MM:SS (no offset)

include

Specify additional fields to return per collector. Can be one or more of: type, status, response_count, date_created, date_modified, url. Optionally, specify "everything" to return all possible fields.

all_pages

Return all pages of respondents? Default is TRUE, which will fetch all responses (and cause n/100 calls to the API).

oauth_token

Your OAuth 2.0 token. By default, retrieved from get_token().

Value

a data.frame (technically a tibble) with each collector and its information.